Simplify and optimize commit clean operation.

+++ bfile/mfile.c
@@ -1,4 +1,4 @@
-/* $Id: mfile.c,v 1.52.2.1 2006-08-14 10:38:50 adam Exp $
+/* $Id: mfile.c,v 1.52.2.2 2006-10-04 09:07:19 adam Exp $
    Copyright (C) 1995,1996,1997,1998,1999,2000,2001,2002,2003
    Index Data Aps
 
@@ -244,7 +244,6 @@ MFile_area mf_init(const char *name, const char *spec, const char *base)
 	    	meta_f->next = ma->mfiles;
 	    	meta_f->open = 0;
 	    	meta_f->cur_file = -1;
-                meta_f->unlink_flag = 0;
 	    	ma->mfiles = meta_f;
 	    	strcpy(meta_f->name, metaname);
 	    	part_f = &meta_f->files[0];
@@ -301,24 +300,11 @@ void mf_destroy(MFile_area ma)
 	dp = dp->next;
 	xfree (d);
     }
-    meta_f = ma->mfiles;
-    while (meta_f)
-    {
-	int i;
-	meta_file *m = meta_f;
-	
-	for (i = 0; i<m->no_files; i++)
-	{
-	    xfree (m->files[i].path);
-	}
-	zebra_mutex_destroy (&meta_f->mutex);
-	meta_f = meta_f->next;
-	xfree (m);
-    }
+    mf_reset(ma, 0);
     xfree (ma);
 }
 
-void mf_reset(MFile_area ma)
+void mf_reset(MFile_area ma, int unlink_flag)
 {
     meta_file *meta_f;
 
@@ -330,13 +316,16 @@ void mf_reset(MFile_area ma)
 	int i;
 	meta_file *m = meta_f;
 
+	meta_f = meta_f->next;
+
 	assert (!m->open);
 	for (i = 0; i<m->no_files; i++)
 	{
-	    unlink (m->files[i].path);
+            if (unlink_flag)
+                unlink (m->files[i].path);
 	    xfree (m->files[i].path);
 	}
-	meta_f = meta_f->next;
+	zebra_mutex_destroy (&m->mutex);
 	xfree (m);
     }
     ma->mfiles = 0;
@@ -344,7 +333,6 @@ void mf_reset(MFile_area ma)
 
 /*
  * Open a metafile.
- * If !ma, Use MF_DEFAULT_AREA.
  */
 MFile mf_open(MFile_area ma, const char *name, int block_size, int wflag)
 {
@@ -376,7 +364,6 @@ MFile mf_open(MFile_area ma, const char *name, int block_size, int wflag)
     	mnew->files[0].top = -1;
     	mnew->files[0].number = 0;
     	mnew->files[0].fd = -1;
-        mnew->unlink_flag = 0;
     	mnew->min_bytes_creat = MF_MIN_BLOCKS_CREAT * block_size;
     	for (dp = ma->dirs; dp && dp->max_bytes >= 0 && dp->avail_bytes <
 	    mnew->min_bytes_creat; dp = dp->next);
@@ -441,8 +428,6 @@ int mf_close(MFile mf)
     	    close(mf->files[i].fd);
     	    mf->files[i].fd = -1;
 	}
-        if (mf->unlink_flag)
-            unlink(mf->files[i].path);
     }
     mf->open = 0;
     return 0;
@@ -587,27 +572,3 @@ int mf_write(MFile mf, int no, int offset, int nbytes, const void *buf)
     return 0;
 }
 
-/*
- * Destroy a metafile, unlinking component files. File must be open.
- */
-int mf_unlink(MFile mf)
-{
-    if (mf->open)
-        mf->unlink_flag = 1;
-    else
-    {
-        int i;
-        for (i = 0; i<mf->no_files; i++)
-            unlink(mf->files[i].path);
-    }
-    return 0;
-}
-
-/*
- * Unlink the file by name, rather than MFile-handle. File should be closed.
- */
-int mf_unlink_name(MFile_area ma, const char *name)
-{
-    abort();
-    return 0;
-}
